refactor(sync): fold pending-sync claims into a keyed record store (stage-4)
Extract src/sync/pending-sync-store.ts: one PendingSyncRecord per retained
key (claimant peers/set/indexes, round-robin cursor, expiry deadline)
replaces the former lockstep map family — syncInFlightQueueClaimants,
syncInFlightQueueClaimantIndexes, syncInFlightQueueRoundRobinCursor and
pendingSyncKeyExpiryNodes fold into the records map. The expiry heap now
holds the records themselves, unioned with the admission-reservation
expiry nodes; heap staleness checks become record-identity checks.

Preserved contracts:
- the public map instances (syncInFlightQueue, syncInFlightQueueInverted,
  syncInFlight, expiry/alias maps, admission accounting) survive as the
  store's physical indexes — the SAME instances, exposed on the
  synchronizer under their historical names via accessors
- directly seeded public maps hydrate into records on first access with
  identical claim-count effects (pinned)
- one add path (addPendingSyncClaim), unified removal (clearSyncProcessKey
  / removePendingSyncClaim / TTL expiry / close), swap-remove cursor
  modulo semantics and alias-transplant deadlines byte-equivalent (pinned)
- admission machinery moved verbatim; the dispatch-epoch check stays a
  caller-supplied predicate (late-bound closure, sinon-safe)
- the synchronizer keeps same-named delegators for every method tests or
  remaining code call, so instance-level stubs and spies keep working

The moved *Admission*/claim state keeps its ratchet coverage via a new
fence-ratchet TARGETS entry for the store module.
